🔵 partage de omit() dans common/
